InterfaceTracker: Remove redundant interface names
Listing e.g. "wlan0+" and "wlan+" is redundant, because "wlan+" matches
any name starting with "wlan".  There are several of these, and
eliminating them reduces the number of rules that need to be applied.

Also, move the wifi rules up to the front of the "afwall" chain.  It's
a shorter list, and wifi is likely to handle higher data rates.