Refactor fixLeak code
Delete the startup script when the user unchecks the fixLeak box.  Run the
root copy/delete commands in an AsyncTask instead of blocking the UI
thread.  Make sure the checkbox status reflects whether the script is
installed on the filesystem.  Gray out the option if the init.d directory
cannot be found.  Improve error reporting and consolidate redundant code.