First attempt at fixing NFLOG log output
When using the LOG target, the kernel collects the logs in the background
and dumps them out when we run "dmesg".

But when using the NFLOG target, the "nflog" utility needs to be running
to collect the data in real time.  This isn't currently happening.

This patch starts up a new background shell + service in order to collect
the NFLOG data.  Since the usage of nflog is quite different from dmesg,
Api and LogActivity were modified accordingly.

Note that nflog does NOT currently terminate when AFWall exits.  And only
one instance of nflog can be active at a time.  In a future commit we
should make nflog more sensitive to kills (e.g. terminate when stdin gets
closed).