Api: Install new multiarch binaries
Check Build.CPU_ABI to determine whether to install ARM, x86, or MIPS
binaries for busybox/iptables/nflog.  Also, store a new "appVersion"
shared preference so that we know when the app has been upgraded.
On each upgrade (and only after an upgrade) we will check to see if
the native binaries need to be updated.