libsuperuser: Missing callbacks if shell dies in Interactive mode
If the shell dies while a command is executing, and the watchdog timer
is enabled, the onCommandResult() callback will get invoked so that the
caller can handle the error.  But if the shell dies between commands,
the queued commands will get stuck in limbo and the caller will never
receive a status indication.  This will probably cause the app to freeze
until the user kills it.

The easiest way to reproduce the problem is to try running
libsuperuser_example in interactive mode on the SDK emulator (which has
a trivial "su" implementation that denies root access to all apps).  It
will hang at "Requesting root privilege" forever.