Update afwallstart
Removed useless installed or not check, for what? This can be done under the GUI or via packageManager 
Logging option for debugging reasons added, normal users won't recognize this, so everything is okay (own doesn't need logging because this will show only AFL related stuff if the OS is already booted)
DO NOT allow holes! To ensure that nothing goes trough until AFWall+ was successful started we flush all rules (should be re-applied under AFWall+ after Android was finally started)
Ensure that port forwarding is enabled, this is not persistent, until the script is removed the kernel or build.prop/userinit will be used again - PF is needed for e.g. NAT, if not you get error msg's (this is mostly for custom users and some pre-defined rules we use)


The only problem I found is that we don't normally need a check for the built-in tables, AOSP already includes iptables so we should detect if iptables are present and installing them or not, currently this is broken in AFWall+ because the binaries are always extracted to /data/data/dev.ukanth.ufirewall/app_bin - one goal should be to extract and check them in this script if they are really present and necessary.