Implement a service to monitor for network connectivity changes.
As of Android 7.0 (API level 24), it is no longer sufficient
to declare a broadcast receiver in the app manifest to receive
the CONNECTIVITY_ACTION broadcast[1].  This commit creates a
service that in turn creates the system broadcast receiver needed
to get the connectivity change broadcasts.

This is an attempt to resolve issue #709[2].  Tested and working
on Android 6.0.1 and 7.1.2.

[1] https://developer.android.com/reference/android/net/ConnectivityManager.html#CONNECTIVITY_ACTION
[2] https://github.com/ukanth/afwall/issues/709