Fix #650: Use root permissions to check if paths exist
For example the paths used with Magisk are not accessible by normal users. This
commit fixes the issue by using root permissions to test the paths.

Those paths are not accessible by normal users, because for example they contain
apps' private data, or Magisk is trying to hide itself from detection.